updated issue_type for import error creation to a text choice
