+++ taric_parsers/importer.py
@@ -6,6 +6,7 @@ from django.db import IntegrityError
 from django.db import transaction
 
 from importer.models import BatchImportError
+from importer.models import BatchImportErrorIssueType
 from importer.models import ImportBatch
 from taric.models import Envelope
 from taric_parsers.parsers.additional_code_parsers import *  # noqa
@@ -199,7 +200,7 @@ class TaricImporter:
 
             transaction_order += 1
 
-        if len(self.issues("ERROR")) > 0:
+        if len(self.issues(BatchImportErrorIssueType.ERROR)) > 0:
             transaction.set_rollback(True)
 
     def commit_changes_from_message(
@@ -256,7 +257,7 @@ class TaricImporter:
                         message,
                         related_tag="self",
                         related_identity_keys=message.taric_object.model_query_parameters(),
-                        issue_type="ERROR",
+                        issue_type=BatchImportErrorIssueType.ERROR,
                         message=msg,
                     )
 
@@ -288,9 +289,9 @@ class TaricImporter:
             - COMPLETED_WITH_WARNINGS : Successful import, but with warnings
             - FAILED : failed import, no data committed to the database. Issues recorded against the import.
         """
-        if len(self.issues("ERROR")) > 0:
+        if len(self.issues(BatchImportErrorIssueType.ERROR)) > 0:
             return "FAILED"
-        elif len(self.issues("WARNING")) > 0:
+        elif len(self.issues(BatchImportErrorIssueType.WARNING)) > 0:
             return "COMPLETED_WITH_WARNINGS"
         else:
             return "COMPLETED"
@@ -396,7 +397,7 @@ class TaricImporter:
         related_tag="",
         related_identity_keys=None,
         message="",
-        issue_type="ERROR",
+        issue_type=BatchImportErrorIssueType.ERROR,
     ):
         """
         Creates an import issue that will be recorded against the database at
@@ -407,6 +408,7 @@ class TaricImporter:
             related_tag: (optional) str, the XML tag for the parsed message
             related_identity_keys: (optional) dict, a dictionary of keys defining the identity of the object.
             message: (optional) str, A string describing the encountered issue / warning.
+            issue_type: (optional) str, from choices BatchImportErrorIssueType.choices (ERROR, WARNING and INFORMATION)
         """
 
         if related_identity_keys is None:
@@ -574,10 +576,10 @@ class TaricImporter:
                 parsed_message.taric_object.import_changes = False
 
                 msg = f"Children of Taric objects of type {parsed_message.taric_object.__class__.model.__name__} can't be deleted directly. This change will not be imported."
-                issue_type = "WARNING"
+                issue_type = BatchImportErrorIssueType.WARNING
             else:
                 msg = f"Taric objects of type {parsed_message.taric_object.__class__.model.__name__} can't be deleted"
-                issue_type = "ERROR"
+                issue_type = BatchImportErrorIssueType.ERROR
 
             self.create_import_issue(
                 parsed_message,
