+++ taric_parsers/tasks.py
@@ -5,6 +5,7 @@ from typing import Sequence
 import taric_parsers.importer
 from common.celery import app
 from importer.models import BatchImportError
+from importer.models import BatchImportErrorIssueType
 from importer.models import ImportBatch
 from importer.models import ImporterChunkStatus
 from importer.models import ImporterXMLChunk
@@ -56,7 +57,10 @@ def import_chunk(
             workbasket=workbasket,
         )
 
-        if len(importer.issues(filter_by_issue_type="ERROR")) > 0:
+        if (
+            len(importer.issues(filter_by_issue_type=BatchImportErrorIssueType.ERROR))
+            > 0
+        ):
             chunk.status = ImporterChunkStatus.ERRORED
         else:
             chunk.status = ImporterChunkStatus.DONE
@@ -79,7 +83,7 @@ def import_chunk(
             taric_change_type="",
             object_details="",
             transaction_id=0,
-            issue_type="ERROR",
+            issue_type=BatchImportErrorIssueType.ERROR,
         )
 
         import_error.save()
