+++ taric_parsers/tests/test_forms.py
@@ -1,0 +1,3 @@
+import pytest
+
+pytestmark = pytest.mark.django_db
